app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
backpack==0.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.7
cleo==0.6.8
click==8.0.3
cmake==3.25.0
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dateparser==1.1.3
dbus-python==1.2.18
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dlib==19.24.0
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.18.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0
executing==1.2.0
fabric==3.0.0
fake-useragent==0.1.14
Faker==0.8.18
fastapi==0.85.1
fasteners==0.14.1
fastjsonschema==2.16.2
Flask==2.0.2
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
graphene==3.1.1
graphene-pydantic==0.4.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.13.0
htmldate==1.3.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
imutils==0.5.4
inflection==0.3.1
iniconfig==1.1.1
invoke==2.0.0
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.0.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.2
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.8.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mysql-connector-python==8.0.31
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.3
oauthlib==3.2.0
olefile==0.46
openai==0.19.0
opencv-python==4.6.0.66
openpyxl==3.0.9
orator==0.9.9
outcome==1.2.0
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.35
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pastel==0.1.1
pendulum==1.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.12.4
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==16.12.2
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.4.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.0
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-gql==0.2.4
python-jose==3.3.0
python-json-logger==2.0.4
python-multipart==0.0.5
pytz==2021.3
pytzdata==2020.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.3.2
reportlab==3.6.8
requests==2.26.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
schedule==1.1.0
scipy==1.9.3
SecretStorage==3.3.1
selenium==4.6.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.20.4
starlette-graphene3==0.6.0
starlette-graphql==0.2.1
supervisor==4.2.1
systemd-python==234
terminado==0.17.1
text-unidecode==1.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.62.3
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
twilio==7.15.0
typing_extensions==4.4.0
tzlocal==1.5.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unicorn==2.0.0
uri-template==1.2.0
urllib3==1.26.7
usb-creator==0.3.7
user-agent==0.1.10
uvicorn==0.15.0
validators==0.20.0
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.8.4
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.0.2
widgetsnbextension==4.0.4
wrapt==1.14.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.8.2
zipp==1.0.0
